Add files via upload
by default, makeRandomRecombinants.py makes breakpoints happen not in the first or last 5000 nt in the chromosome. this script will produce breakpoints ONLY within the first or last 5000 nt of the chromosome. this is to correct the distribution to ensure that ripples does just as well regardless of where breakpoints happen.